feat: use interm_strg_locn to store Other facility type description
