+++ backend/src/test/java/ca/bc/gov/backendstartapi/service/SeedlotServiceTest.java
@@ -4,41 +4,25 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.when;
 
-import ca.bc.gov.backendstartapi.dto.ParentTreeGeneticQualityDto;
 import ca.bc.gov.backendstartapi.dto.SeedlotApplicationPatchDto;
 import ca.bc.gov.backendstartapi.dto.SeedlotCreateDto;
 import ca.bc.gov.backendstartapi.dto.SeedlotCreateResponseDto;
-import ca.bc.gov.backendstartapi.dto.SeedlotFormCollectionDto;
-import ca.bc.gov.backendstartapi.dto.SeedlotFormExtractionDto;
-import ca.bc.gov.backendstartapi.dto.SeedlotFormInterimDto;
-import ca.bc.gov.backendstartapi.dto.SeedlotFormOrchardDto;
-import ca.bc.gov.backendstartapi.dto.SeedlotFormOwnershipDto;
-import ca.bc.gov.backendstartapi.dto.SeedlotFormParentTreeSmpDto;
-import ca.bc.gov.backendstartapi.dto.SeedlotFormSubmissionDto;
 import ca.bc.gov.backendstartapi.entity.GeneticClassEntity;
 import ca.bc.gov.backendstartapi.entity.SeedlotSourceEntity;
 import ca.bc.gov.backendstartapi.entity.SeedlotStatusEntity;
 import ca.bc.gov.backendstartapi.entity.embeddable.EffectiveDateRange;
 import ca.bc.gov.backendstartapi.entity.seedlot.Seedlot;
-import ca.bc.gov.backendstartapi.exception.ConeCollectionMethodNotFoundException;
 import ca.bc.gov.backendstartapi.exception.InvalidSeedlotRequestException;
-import ca.bc.gov.backendstartapi.exception.MethodOfPaymentNotFoundException;
 import ca.bc.gov.backendstartapi.exception.SeedlotNotFoundException;
-import ca.bc.gov.backendstartapi.exception.SeedlotParentTreeNotFoundException;
 import ca.bc.gov.backendstartapi.exception.SeedlotSourceNotFoundException;
-import ca.bc.gov.backendstartapi.exception.SmpMixNotFoundException;
 import ca.bc.gov.backendstartapi.repository.GeneticClassRepository;
 import ca.bc.gov.backendstartapi.repository.SeedlotRepository;
 import ca.bc.gov.backendstartapi.repository.SeedlotSourceRepository;
 import ca.bc.gov.backendstartapi.repository.SeedlotStatusRepository;
 import ca.bc.gov.backendstartapi.security.LoggedUserService;
-import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Assertions;
@@ -96,78 +80,6 @@ class SeedlotServiceTest {
   private SeedlotCreateDto createSeedlotDto() {
     return new SeedlotCreateDto(
         "00012797", "01", "user.lastname@domain.com", "FDI", "TPT", true, true, 'A');
-  }
-
-  private SeedlotFormSubmissionDto mockSeedlotFormDto() {
-    // step 1
-    SeedlotFormCollectionDto collectionDto =
-        new SeedlotFormCollectionDto(
-            "00012797",
-            "02",
-            LocalDateTime.now(),
-            LocalDateTime.now(),
-            new BigDecimal("2"),
-            new BigDecimal("4"),
-            new BigDecimal("8"),
-            "seedlot comment",
-            List.of(1, 2));
-
-    // step 2
-    SeedlotFormOwnershipDto ownerDto =
-        new SeedlotFormOwnershipDto(
-            "00012797",
-            "02",
-            new BigDecimal("100"),
-            new BigDecimal("100"),
-            new BigDecimal("5"),
-            "CLA",
-            "ITC");
-
-    // step 3
-    SeedlotFormInterimDto interimDto =
-        new SeedlotFormInterimDto(
-            "00012797", "02", LocalDateTime.now(), LocalDateTime.now(), "Some location", "OCV");
-
-    // step 4
-    SeedlotFormOrchardDto orchardDto =
-        new SeedlotFormOrchardDto(
-            List.of("405"), "F3", "M3", false, true, false, 22, new BigDecimal("45.6"), "true");
-
-    // step 5
-    ParentTreeGeneticQualityDto ptgqDto =
-        new ParentTreeGeneticQualityDto("BV", "GVO", new BigDecimal("18"));
-    SeedlotFormParentTreeSmpDto parentTreeDto =
-        new SeedlotFormParentTreeSmpDto(
-            "87",
-            4023,
-            "87",
-            new BigDecimal("1"),
-            new BigDecimal("5"),
-            6,
-            2,
-            50,
-            new BigDecimal("100"),
-            List.of(ptgqDto));
-
-    // step 6
-    SeedlotFormExtractionDto extractionDto =
-        new SeedlotFormExtractionDto(
-            "00012797",
-            "02",
-            LocalDateTime.now(),
-            LocalDateTime.now(),
-            "00012797",
-            "02",
-            LocalDateTime.now(),
-            LocalDateTime.now());
-
-    return new SeedlotFormSubmissionDto(
-        collectionDto,
-        List.of(ownerDto),
-        interimDto,
-        orchardDto,
-        List.of(parentTreeDto),
-        extractionDto);
   }
 
   @BeforeEach
@@ -437,171 +349,5 @@ class SeedlotServiceTest {
         testDto.seedlotSourceCode(), patchedSeedlot.getSeedlotSource().getSeedlotSourceCode());
     assertEquals(testDto.toBeRegistrdInd(), patchedSeedlot.getIntendedForCrownLand());
     assertEquals(testDto.bcSourceInd(), patchedSeedlot.getSourceInBc());
-  }
-
-  @Test
-  @DisplayName("Seedlot form submit - Seedlot not found")
-  void submitSeedlotForm_seedlotNotFound_shouldFail() {
-    when(seedlotRepository.findById("5432")).thenThrow(new SeedlotNotFoundException());
-
-    Assertions.assertThrows(
-        SeedlotNotFoundException.class,
-        () -> {
-          seedlotService.submitSeedlotForm("5432", mockSeedlotFormDto());
-        });
-  }
-
-  @Test
-  @DisplayName("Seedlot form submit - Cone Collection Method not found")
-  void submitSeedlotForm_coneCollectionMethodNotFound_shouldFail() {
-    Seedlot seedlot = new Seedlot("5432");
-    when(seedlotRepository.findById("5432")).thenReturn(Optional.of(seedlot));
-
-    doThrow(new ConeCollectionMethodNotFoundException())
-        .when(seedlotCollectionMethodService)
-        .saveSeedlotFormStep1(any(), any());
-
-    Assertions.assertThrows(
-        ConeCollectionMethodNotFoundException.class,
-        () -> {
-          seedlotService.submitSeedlotForm("5432", mockSeedlotFormDto());
-        });
-  }
-
-  @Test
-  @DisplayName("Seedlot form submit - Method of Payment not found")
-  void submitSeedlotForm_methodOfPaymentNotFound_shouldFail() {
-    Seedlot seedlot = new Seedlot("5432");
-    when(seedlotRepository.findById("5432")).thenReturn(Optional.of(seedlot));
-
-    doNothing().when(seedlotCollectionMethodService).saveSeedlotFormStep1(any(), any());
-
-    doThrow(new MethodOfPaymentNotFoundException())
-        .when(seedlotOwnerQuantityService)
-        .saveSeedlotFormStep2(any(), any());
-
-    Assertions.assertThrows(
-        MethodOfPaymentNotFoundException.class,
-        () -> {
-          seedlotService.submitSeedlotForm("5432", mockSeedlotFormDto());
-        });
-  }
-
-  @Test
-  @DisplayName("Seedlot form submit - Seedlot Parent Tree not found")
-  void submitSeedlotForm_seedlotParentTreeNotFound_shouldFail() {
-    Seedlot seedlot = new Seedlot("5432");
-    when(seedlotRepository.findById("5432")).thenReturn(Optional.of(seedlot));
-
-    doNothing().when(seedlotCollectionMethodService).saveSeedlotFormStep1(any(), any());
-    when(seedlotOwnerQuantityService.saveSeedlotFormStep2(any(), any())).thenReturn(List.of());
-    doNothing().when(seedlotOrchardService).saveSeedlotFormStep4(any(), any());
-    when(seedlotParentTreeService.saveSeedlotFormStep5(any(), any())).thenReturn(List.of());
-
-    doThrow(new SeedlotParentTreeNotFoundException())
-        .when(seedlotParentTreeGeneticQualityService)
-        .saveSeedlotFormStep5(any(), any());
-
-    Assertions.assertThrows(
-        SeedlotParentTreeNotFoundException.class,
-        () -> {
-          seedlotService.submitSeedlotForm("5432", mockSeedlotFormDto());
-        });
-  }
-
-  @Test
-  @DisplayName("Seedlot form submit - Smp Mix not found")
-  void submitSeedlotForm_smpMixNotFound_shouldFail() {
-    Seedlot seedlot = new Seedlot("5432");
-    when(seedlotRepository.findById("5432")).thenReturn(Optional.of(seedlot));
-
-    doNothing().when(seedlotCollectionMethodService).saveSeedlotFormStep1(any(), any());
-    when(seedlotOwnerQuantityService.saveSeedlotFormStep2(any(), any())).thenReturn(List.of());
-    doNothing().when(seedlotOrchardService).saveSeedlotFormStep4(any(), any());
-    when(seedlotParentTreeService.saveSeedlotFormStep5(any(), any())).thenReturn(List.of());
-    doNothing().when(seedlotParentTreeGeneticQualityService).saveSeedlotFormStep5(any(), any());
-    when(seedlotGeneticWorthService.saveSeedlotFormStep5(any(), any())).thenReturn(List.of());
-    when(smpMixService.saveSeedlotFormStep5(any(), any())).thenReturn(List.of());
-
-    doThrow(new SmpMixNotFoundException())
-        .when(smpMixGeneticQualityService)
-        .saveSeedlotFormStep5(any(), any());
-
-    Assertions.assertThrows(
-        SmpMixNotFoundException.class,
-        () -> {
-          seedlotService.submitSeedlotForm("5432", mockSeedlotFormDto());
-        });
-  }
-
-  @Test
-  @DisplayName("Seedlot form submit - Success")
-  void submitSeedlotForm_happyPath_shouldSucceed() {
-    Seedlot seedlot = new Seedlot("5432");
-    when(seedlotRepository.findById("5432")).thenReturn(Optional.of(seedlot));
-
-    doNothing().when(seedlotCollectionMethodService).saveSeedlotFormStep1(any(), any());
-    when(seedlotOwnerQuantityService.saveSeedlotFormStep2(any(), any())).thenReturn(List.of());
-    doNothing().when(seedlotOrchardService).saveSeedlotFormStep4(any(), any());
-    when(seedlotParentTreeService.saveSeedlotFormStep5(any(), any())).thenReturn(List.of());
-    doNothing().when(seedlotParentTreeGeneticQualityService).saveSeedlotFormStep5(any(), any());
-    when(seedlotGeneticWorthService.saveSeedlotFormStep5(any(), any())).thenReturn(List.of());
-    when(smpMixService.saveSeedlotFormStep5(any(), any())).thenReturn(List.of());
-    doNothing().when(smpMixGeneticQualityService).saveSeedlotFormStep5(any(), any());
-    doNothing().when(seedlotParentTreeSmpMixService).saveSeedlotFormStep5(any(), any());
-
-    SeedlotStatusEntity ssEntity = new SeedlotStatusEntity();
-    ssEntity.setSeedlotStatusCode("SUB");
-    when(seedlotStatusService.getValidSeedlotStatus(any())).thenReturn(Optional.of(ssEntity));
-
-    SeedlotFormSubmissionDto mockedForm = mockSeedlotFormDto();
-    SeedlotCreateResponseDto scDto = seedlotService.submitSeedlotForm("5432", mockedForm);
-
-    Assertions.assertNotNull(scDto);
-    Assertions.assertEquals("5432", scDto.seedlotNumber());
-    Assertions.assertEquals("SUB", scDto.seedlotStatusCode());
-
-    Assertions.assertEquals(
-        mockedForm.seedlotFormInterimDto().intermStrgClientNumber(),
-        seedlot.getInterimStorageClientNumber());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormInterimDto().intermStrgLocnCode(),
-        seedlot.getInterimStorageLocationCode());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormInterimDto().intermStrgStDate(),
-        seedlot.getInterimStorageStartDate());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormInterimDto().intermStrgEndDate(),
-        seedlot.getInterimStorageEndDate());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormInterimDto().intermStrgLocn(),
-        seedlot.getInterimStorageLocationDescription());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormInterimDto().intermFacilityCode(),
-        seedlot.getInterimStorageFacilityCode());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().extractoryClientNumber(),
-        seedlot.getExtractionClientNumber());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().extractoryLocnCode(),
-        seedlot.getExtractionLocationCode());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().extractionStDate(),
-        seedlot.getExtractionStartDate());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().extractionEndDate(),
-        seedlot.getExtractionEndDate());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().storageClientNumber(),
-        seedlot.getStorageClientNumber());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().storageLocnCode(),
-        seedlot.getStorageLocationCode());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().temporaryStrgStartDate(),
-        seedlot.getTemporaryStorageStartDate());
-    Assertions.assertEquals(
-        mockedForm.seedlotFormExtractionDto().temporaryStrgEndDate(),
-        seedlot.getTemporaryStorageEndDate());
   }
 }
